+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.24.0] - 2024-05-22
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): User list at `manage` page.
+
 ### Changed
 - [Marek Kurańda](https://github.com/mjkuranda): `Manage` bookmark is not available for workers.
 - [Marek Kurańda](https://github.com/mjkuranda): Settings subpage to Manage and settings page content moved to `/manage` page.